refactor: #25 Update project structure
