+++ src/event-handling/setup-event-listener.ts
@@ -1,6 +1,6 @@
-import setUpPlanetEventListeners from '../modules/map/input/event-handling/planet-event-listener'
-import setUpRobotEventListeners from '../modules/robot/input/event-handling/robot-event-listener'
-import setUpTradingEventListeners from '../modules/trading/bank-account/input/event-handling/bankAccount-event-listener'
+import setUpPlanetEventListeners from '../modules/map/adapters/input/event-handling/planet-event-listener'
+import setUpRobotEventListeners from '../modules/robot/adapters/input/event-handling/robot-event-listener'
+import setUpTradingEventListeners from '../modules/trading/bank-account/adapters/input/event-handling/bankAccount-event-listener'
 
 export default function setUpEventListeners() {
   setUpTradingEventListeners()
