+++ src/modules/robot/adapters/output/commands/index.ts
@@ -1,0 +1,14 @@
+import { buyRobots } from './buyRobots'
+import { engangeBattle } from './engangeBattle'
+import { mineResources } from './mineResources'
+import { moveRobot } from './moveRobot'
+
+const robotCommandService = {
+  buyRobots,
+  engangeBattle,
+  mineResources,
+  moveRobot,
+}
+
+export default robotCommandService
+export { buyRobots, engangeBattle, mineResources, moveRobot }
