+++ src/modules/robot/adapters/output/commands/moveRobot.ts
@@ -1,0 +1,6 @@
+import { MoveCommandData, MoveCommand } from '../../../../../shared/commands'
+import { sendCommand } from '../../../../../shared/sendCommand'
+
+export async function moveRobot({ robotId, planetId }: MoveCommandData) {
+  return sendCommand<MoveCommand>({ type: 'movement', data: { robotId, planetId } })
+}
