--- src/modules/trading/bank-account/output/commands/index.ts
@@ -1,10 +1,0 @@
-import { buyUpgrade } from './buyUpgrade'
-import { sellRessources } from './sellRessources'
-
-const tradingCommandsService = {
-  buyUpgrade,
-  sellRessources,
-}
-
-export default tradingCommandsService
-export { buyUpgrade, sellRessources }
