--- src/modules/trading/bank-account/output/data-access/index.ts
@@ -1,19 +1,0 @@
-import { BankAccount } from '../../domain/model/bankAccount'
-import { BankAccountDb } from '../../domain/use-cases/types'
-
-// TODO: Provide further logic
-let bankAccountData: BankAccount | null = null
-
-const bankAccountDb: BankAccountDb = {
-  insert: (bankAccount: BankAccount) => {
-    bankAccountData = bankAccount
-    return bankAccount
-  },
-  find: () => bankAccountData as BankAccount,
-  update: (bankAccount: BankAccount) => {
-    bankAccountData = bankAccount
-    return bankAccount
-  },
-}
-
-export default bankAccountDb
