+++ src/modules/trading/inventory/adapters/output/commands/sellRessources.ts
@@ -1,0 +1,6 @@
+import { SellTradablesCommandData, SellTradablesCommand } from '../../../../../../shared/commands'
+import { sendCommand } from '../../../../../../shared/sendCommand'
+
+export async function sellRessources({ robotId }: SellTradablesCommandData) {
+  return sendCommand<SellTradablesCommand>({ type: 'selling', data: { robotId } })
+}
