+++ src/modules/trading/inventory/adapters/output/data-access/index.ts
@@ -1,0 +1,12 @@
+import { Inventory } from '../../../domain/model/inventory'
+import { InventoryDb } from '../../../domain/use-cases/types'
+
+/* eslint-disable @typescript-eslint/no-unused-vars */
+const inventoryDb: InventoryDb = {
+  find: (name: string) => ({}) as Inventory,
+  findAll: () => [] as Array<Inventory>,
+  insert: (inventory: Inventory) => ({}) as Inventory,
+  update: (inventory: Inventory) => ({}) as Inventory,
+}
+
+export default inventoryDb
