+++ src/modules/trading/resource/adapters/output/data-access/index.ts
@@ -1,0 +1,11 @@
+import { ResourceDb } from '../../../domain/use-cases/types'
+import { Resource } from '../../../domain/model/resource'
+
+// TODO: Provide further logic
+const resourceDb: ResourceDb = {
+  insert: () => ({}) as Resource,
+  find: () => ({}) as Resource,
+  update: () => ({}) as Resource,
+}
+
+export default resourceDb
